requirements.txt
diff-match-patch==20121119
Django==2.1.2
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.7.2
django-datatables-view==1.17.0
-e git+https://github.com/django-import-export/django-import-export.git@54232db4a511cc5f0b085c6bb192c3619f27e452#egg=django_import_export
django-pyodbc-azure==2.1.0.0
django-python3-ldap==0.11.1
et-xmlfile==1.0.1
jdcal==1.4
ldap3==2.3
odfpy==1.3.6
openpyxl==2.5.8
pyasn1==0.3.2
pyodbc==4.0.24
pytz==2018.5
PyYAML==3.13
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
